+++ regression/automatic_regression.py
@@ -73,6 +73,7 @@ modules = [
     'scripts/rocket_network/Rocketdyne_J2.py',   
     'scripts/sizing_loop/sizing_loop.py',
     'scripts/solar_network/solar_network.py',
+    'scripts/optimization_packages/optimization_packages.py'
     'scripts/solar_network/solar_low_fidelity_network.py',
     'scripts/solar_radiation/solar_radiation.py',
     'scripts/SU2_surrogate/BWB-450.py',
