+++ regression/scripts/optimization_packages/vehicle_opt_pack.py
@@ -1,0 +1,66 @@
+# Vehicles.py
+# 
+# Created: Sep. 2019, M. Clarke
+# ----------------------------------------------------------------------        
+#   Imports
+# ----------------------------------------------------------------------    
+
+import SUAVE
+
+# ----------------------------------------------------------------------
+#   Define the Vehicle
+# ----------------------------------------------------------------------
+
+def setup():
+    
+    base_vehicle = base_setup()
+    configs = configs_setup(base_vehicle)
+    
+    return configs
+    
+    
+def base_setup():
+
+    # ------------------------------------------------------------------
+    #   Initialize the Vehicle
+    # ------------------------------------------------------------------    
+
+    vehicle = SUAVE.Vehicle()
+    vehicle.tag = 'Symbolic_Vehicle'    
+
+
+    # ------------------------------------------------------------------
+    #   Vehicle-level Properties
+    # ------------------------------------------------------------------    
+
+    # mass properties
+    vehicle.x1 = 0
+    vehicle.x2 = 0
+
+    # ------------------------------------------------------------------
+    #   Vehicle Definition Complete
+    # ------------------------------------------------------------------
+
+    return vehicle
+
+
+
+
+# ----------------------------------------------------------------------
+#   Define the Configurations
+# ---------------------------------------------------------------------
+
+def configs_setup(vehicle):
+    
+    # ------------------------------------------------------------------
+    #   Initialize Configurations
+    # ------------------------------------------------------------------
+
+    configs = SUAVE.Components.Configs.Config.Container()
+
+    base_config = SUAVE.Components.Configs.Config(vehicle)
+    base_config.tag = 'base'
+    configs.append(base_config)
+
+    # done!
+    return configs
